extract html .as_tantivy into textfield trait
