+++ crates/core/src/webpage/html/into_tantivy.rs
@@ -199,277 +199,10 @@ impl Html {
         {
             let tantivy_field = schema
                 .get_field(field.name())
-                .unwrap_or_else(|_| panic!("Unknown field: {}", field.name()));
+                .unwrap_or_else(|_| unreachable!("Unknown field: {}", field.name()));
 
             match field {
-                Field::Text(TextFieldEnum::Title(_)) => doc.add_pre_tokenized_text(
-                    tantivy_field,
-                    cache
-                        .pretokenize_title()
-                        .as_ref()
-                        .map(Clone::clone)
-                        .map_err(|e| anyhow::anyhow!("{}", e))?,
-                ),
-                Field::Text(TextFieldEnum::StemmedTitle(_)) => {
-                    let title = cache
-                        .pretokenize_title()
-                        .as_ref()
-                        .map(Clone::clone)
-                        .map_err(|e| anyhow::anyhow!("{}", e))?;
-                    let mut tokens = title.tokens.clone();
-                    stem_tokens(&mut tokens, self.lang.unwrap_or(Lang::Eng));
-
-                    doc.add_pre_tokenized_text(
-                        tantivy_field,
-                        PreTokenizedString {
-                            text: title.text.clone(),
-                            tokens,
-                        },
-                    );
-                }
-                Field::Text(TextFieldEnum::CleanBody(_)) => doc
-                    .add_pre_tokenized_text(tantivy_field, cache.pretokenize_clean_text().clone()),
-                Field::Text(TextFieldEnum::StemmedCleanBody(_)) => {
-                    let clean_text = cache.pretokenize_clean_text();
-                    let mut tokens = clean_text.tokens.clone();
-                    stem_tokens(&mut tokens, self.lang.unwrap_or(Lang::Eng));
-
-                    doc.add_pre_tokenized_text(
-                        tantivy_field,
-                        PreTokenizedString {
-                            text: clean_text.text.clone(),
-                            tokens,
-                        },
-                    );
-                }
-                Field::Text(TextFieldEnum::CleanBodyBigrams(_)) => {
-                    doc.add_text(
-                        tantivy_field,
-                        self.clean_text().cloned().unwrap_or_default(),
-                    );
-                }
-                Field::Text(TextFieldEnum::CleanBodyTrigrams(_)) => {
-                    doc.add_text(
-                        tantivy_field,
-                        self.clean_text().cloned().unwrap_or_default(),
-                    );
-                }
-                Field::Text(TextFieldEnum::TitleBigrams(_)) => {
-                    let title = cache
-                        .pretokenize_title()
-                        .as_ref()
-                        .map(Clone::clone)
-                        .map_err(|e| anyhow::anyhow!("{}", e))?;
-                    doc.add_text(tantivy_field, title.text.clone());
-                }
-                Field::Text(TextFieldEnum::TitleTrigrams(_)) => {
-                    let title = cache
-                        .pretokenize_title()
-                        .as_ref()
-                        .map(Clone::clone)
-                        .map_err(|e| anyhow::anyhow!("{}", e))?;
-                    doc.add_text(tantivy_field, title.text.clone());
-                }
-                Field::Text(TextFieldEnum::Description(_)) => {
-                    let description = cache.pretokenize_description();
-                    doc.add_pre_tokenized_text(tantivy_field, description.clone());
-                }
-                Field::Text(TextFieldEnum::Url(_)) => {
-                    let url = cache.pretokenize_url();
-                    doc.add_pre_tokenized_text(tantivy_field, url.clone())
-                }
-                Field::Text(TextFieldEnum::UrlForSiteOperator(_)) => doc.add_pre_tokenized_text(
-                    tantivy_field,
-                    cache.pretokenize_url_for_site_operator().clone(),
-                ),
-                Field::Text(TextFieldEnum::UrlNoTokenizer(_)) => {
-                    let url = self.url().to_string();
-
-                    doc.add_pre_tokenized_text(
-                        tantivy_field,
-                        PreTokenizedString {
-                            text: url.clone(),
-                            tokens: vec![tantivy::tokenizer::Token {
-                                offset_from: 0,
-                                offset_to: url.len(),
-                                position: 0,
-                                text: url,
-                                position_length: 1,
-                            }],
-                        },
-                    );
-                }
-                Field::Text(TextFieldEnum::SiteWithout(_)) => {
-                    doc.add_pre_tokenized_text(tantivy_field, cache.pretokenize_site().clone())
-                }
-                Field::Text(TextFieldEnum::Domain(_)) => {
-                    doc.add_pre_tokenized_text(tantivy_field, cache.pretokenize_domain().clone())
-                }
-                Field::Text(TextFieldEnum::SiteNoTokenizer(_)) => {
-                    let site = cache.pretokenize_site();
-
-                    doc.add_pre_tokenized_text(
-                        tantivy_field,
-                        PreTokenizedString {
-                            text: site.text.clone(),
-                            tokens: vec![tantivy::tokenizer::Token {
-                                offset_from: 0,
-                                offset_to: site.text.len(),
-                                position: 0,
-                                text: site.text.clone(),
-                                position_length: 1,
-                            }],
-                        },
-                    )
-                }
-                Field::Text(TextFieldEnum::SiteIfHomepageNoTokenizer(_)) => {
-                    let site = cache.pretokenize_site();
-
-                    if self.is_homepage() {
-                        doc.add_pre_tokenized_text(
-                            tantivy_field,
-                            PreTokenizedString {
-                                text: site.text.clone(),
-                                tokens: vec![tantivy::tokenizer::Token {
-                                    offset_from: 0,
-                                    offset_to: site.text.len(),
-                                    position: 0,
-                                    text: site.text.clone(),
-                                    position_length: 1,
-                                }],
-                            },
-                        )
-                    } else {
-                        doc.add_text(tantivy_field, "");
-                    }
-                }
-                Field::Text(TextFieldEnum::DomainNoTokenizer(_)) => {
-                    let domain = cache.pretokenize_domain();
-
-                    doc.add_pre_tokenized_text(
-                        tantivy_field,
-                        PreTokenizedString {
-                            text: domain.text.clone(),
-                            tokens: vec![tantivy::tokenizer::Token {
-                                offset_from: 0,
-                                offset_to: domain.text.len(),
-                                position: 0,
-                                text: domain.text.clone(),
-                                position_length: 1,
-                            }],
-                        },
-                    )
-                }
-                Field::Text(TextFieldEnum::TitleIfHomepage(_)) => {
-                    let title = cache
-                        .pretokenize_title()
-                        .as_ref()
-                        .map(Clone::clone)
-                        .map_err(|e| anyhow::anyhow!("{}", e))?;
-
-                    if self.is_homepage() {
-                        doc.add_pre_tokenized_text(tantivy_field, title);
-                    } else {
-                        doc.add_text(tantivy_field, "");
-                    }
-                }
-                Field::Text(TextFieldEnum::DomainIfHomepage(_)) => {
-                    let domain = cache.pretokenize_domain();
-                    if self.is_homepage() {
-                        doc.add_text(tantivy_field, domain.text.clone());
-                    } else {
-                        doc.add_text(tantivy_field, "");
-                    }
-                }
-                Field::Text(TextFieldEnum::DomainNameNoTokenizer(_)) => {
-                    let domain_name = cache.domain_name();
-
-                    doc.add_pre_tokenized_text(
-                        tantivy_field,
-                        PreTokenizedString {
-                            text: domain_name.clone(),
-                            tokens: vec![tantivy::tokenizer::Token {
-                                offset_from: 0,
-                                offset_to: domain_name.len(),
-                                position: 0,
-                                text: domain_name.clone(),
-                                position_length: 1,
-                            }],
-                        },
-                    );
-                }
-                Field::Text(TextFieldEnum::DomainNameIfHomepageNoTokenizer(_)) => {
-                    let domain_name = cache.domain_name();
-
-                    if self.is_homepage() {
-                        doc.add_pre_tokenized_text(
-                            tantivy_field,
-                            PreTokenizedString {
-                                text: domain_name.clone(),
-                                tokens: vec![tantivy::tokenizer::Token {
-                                    offset_from: 0,
-                                    offset_to: domain_name.len(),
-                                    position: 0,
-                                    text: domain_name.clone(),
-                                    position_length: 1,
-                                }],
-                            },
-                        );
-                    } else {
-                        doc.add_text(tantivy_field, "");
-                    }
-                }
-                Field::Text(TextFieldEnum::DomainIfHomepageNoTokenizer(_)) => {
-                    let domain = cache.pretokenize_domain();
-
-                    if self.url().is_homepage() {
-                        doc.add_pre_tokenized_text(
-                            tantivy_field,
-                            PreTokenizedString {
-                                text: domain.text.clone(),
-                                tokens: vec![tantivy::tokenizer::Token {
-                                    offset_from: 0,
-                                    offset_to: domain.text.len(),
-                                    position: 0,
-                                    text: domain.text.clone(),
-                                    position_length: 1,
-                                }],
-                            },
-                        );
-                    } else {
-                        doc.add_text(tantivy_field, "");
-                    }
-                }
-                Field::Text(TextFieldEnum::AllBody(_)) => {
-                    let all_text = cache
-                        .pretokenize_all_text()
-                        .as_ref()
-                        .map(Clone::clone)
-                        .map_err(|e| anyhow::anyhow!("{}", e))?;
-
-                    doc.add_pre_tokenized_text(tantivy_field, all_text.clone())
-                }
-                Field::Text(TextFieldEnum::RecipeFirstIngredientTagId(_)) => {
-                    doc.add_text(
-                        tantivy_field,
-                        cache.first_ingredient_tag_id().cloned().unwrap_or_default(),
-                    );
-                }
-                Field::Text(TextFieldEnum::SchemaOrgJson(_)) => {
-                    doc.add_text(tantivy_field, cache.schema_json());
-                }
-                Field::Text(TextFieldEnum::FlattenedSchemaOrgJson(_)) => {
-                    doc.add_pre_tokenized_text(
-                        tantivy_field,
-                        cache.pretokenized_schema_json().clone(),
-                    );
-                }
-                Field::Text(TextFieldEnum::MicroformatTags(_)) => {
-                    doc.add_pre_tokenized_text(
-                        tantivy_field,
-                        cache.pretokenize_microformats().clone(),
-                    );
-                }
+                Field::Text(f) => f.add_html_tantivy(self, &mut cache, &mut doc, schema)?,
                 Field::Fast(FastFieldEnum::IsHomepage(_)) => {
                     doc.add_u64(tantivy_field, (self.is_homepage()).into());
                 }
@@ -600,10 +333,7 @@ impl Html {
                         (self.link_density() * FLOAT_SCALING as f64) as u64,
                     );
                 }
-                Field::Text(TextFieldEnum::BacklinkText(_))
-                | Field::Text(TextFieldEnum::SafetyClassification(_))
-                | Field::Text(TextFieldEnum::InsertionTimestamp(_))
-                | Field::Fast(FastFieldEnum::HostCentrality(_))
+                Field::Fast(FastFieldEnum::HostCentrality(_))
                 | Field::Fast(FastFieldEnum::HostCentralityRank(_))
                 | Field::Fast(FastFieldEnum::PageCentrality(_))
                 | Field::Fast(FastFieldEnum::PageCentralityRank(_))
@@ -612,46 +342,10 @@ impl Html {
                 | Field::Fast(FastFieldEnum::Region(_))
                 | Field::Fast(FastFieldEnum::HostNodeID(_))
                 | Field::Fast(FastFieldEnum::TitleEmbeddings(_))
-                | Field::Fast(FastFieldEnum::KeywordEmbeddings(_))
-                | Field::Text(TextFieldEnum::Keywords(_))
-                | Field::Text(TextFieldEnum::DmozDescription(_)) => {}
+                | Field::Fast(FastFieldEnum::KeywordEmbeddings(_)) => {}
             }
         }
 
         Ok(doc)
-    }
-}
-
-fn stemmer_from_lang(lang: &Lang) -> rust_stemmers::Stemmer {
-    match lang {
-        Lang::Ara => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Arabic),
-        Lang::Dan => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Danish),
-        Lang::Nld => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Dutch),
-        Lang::Fin => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Finnish),
-        Lang::Fra => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::French),
-        Lang::Deu => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::German),
-        Lang::Ell => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Greek),
-        Lang::Hun => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Hungarian),
-        Lang::Ita => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Italian),
-        Lang::Por => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Portuguese),
-        Lang::Ron => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Romanian),
-        Lang::Rus => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Russian),
-        Lang::Spa => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Spanish),
-        Lang::Swe => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Swedish),
-        Lang::Tam => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Tamil),
-        Lang::Tur => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::Turkish),
-        _ => rust_stemmers::Stemmer::create(rust_stemmers::Algorithm::English),
-    }
-}
-
-fn stem_tokens(tokens: &mut [tantivy::tokenizer::Token], lang: Lang) {
-    let stemmer = stemmer_from_lang(&lang);
-    for token in tokens {
-        // TODO remove allocation
-        if let Ok(stemmed_str) = std::panic::catch_unwind(|| stemmer.stem(&token.text).into_owned())
-        {
-            token.text.clear();
-            token.text.push_str(&stemmed_str);
-        }
     }
 }
